+++ src/audio_metadata/formats/id3v2.py
@@ -12,13 +12,50 @@ from bidict import frozenbidict
 
 from .id3v2_frames import *
 from .models import Tags
+from .tables import ID3Version
 from ..exceptions import InvalidFrame, InvalidHeader
 from ..structures import DictMixin
 from ..utils import DataReader, decode_synchsafe_int
 
 
 class ID3v2Frames(Tags):
-	FIELD_MAP = frozenbidict({
+	_v22_FIELD_MAP = frozenbidict({
+		'album': 'TAL',
+		'albumartist': 'TP2',
+		'arranger': 'TP4',
+		'artist': 'TP1',
+		'audiodelay': 'TDY',
+		'audiolength': 'TLE',
+		'audiosize': 'TSI',
+		'bpm': 'TBP',
+		'comment': 'COM',
+		'composer': 'TCM',
+		'conductor': 'TP3',
+		'copyright': 'TCR',
+		'date': 'TYE',
+		'discnumber': 'TPA',
+		'encodedby': 'TEN',
+		'encodingsettings': 'TSS',
+		'genre': 'TCO',
+		'grouping': 'TT1',
+		'isrc': 'TRC',
+		'language': 'TLA',
+		'lyricist': 'TXT',
+		'lyrics': 'ULT',
+		'media': 'TMT',
+		'originalalbum': 'TOT',
+		'originalartist': 'TOA',
+		'originalauthor': 'TOL',
+		'originalyear': 'TOR',
+		'pictures': 'PIC',
+		'playcount': 'CNT',
+		'publisher': 'TPB',
+		'subtitle': 'TT3',
+		'title': 'TT2',
+		'tracknumber': 'TRK'
+	})
+
+	_v23_FIELD_MAP = frozenbidict({
 		'album': 'TALB',
 		'albumsort': 'TSOA',
 		'albumartist': 'TPE2',
@@ -29,13 +66,56 @@ class ID3v2Frames(Tags):
 		'audiodelay': 'TDLY',
 		'audiolength': 'TLEN',
 		'audiosize': 'TSIZ',
-		'bpm': 'TBPM', 'comment': 'COMM',
+		'bpm': 'TBPM',
+		'comment': 'COMM',
 		'compilation': 'TCMP',
 		'composer': 'TCOM',
 		'composersort': 'TSOC',
 		'conductor': 'TPE3',
 		'copyright': 'TCOP',
 		'date': 'TYER',
+		'discnumber': 'TPOS',
+		'encodedby': 'TENC',
+		'encodingsettings': 'TSSE',
+		'genre': 'TCON',
+		'grouping': 'TIT1',
+		'isrc': 'TSRC',
+		'language': 'TLAN',
+		'lyricist': 'TEXT',
+		'lyrics': 'USLT',
+		'media': 'TMED',
+		'originalalbum': 'TOAL',
+		'originalartist': 'TOPE',
+		'originalauthor': 'TOLY',
+		'originalyear': 'TORY',
+		'pictures': 'APIC',
+		'playcount': 'PCNT',
+		'publisher': 'TPUB',
+		'subtitle': 'TIT3',
+		'title': 'TIT2',
+		'titlesort': 'TSOT',
+		'tracknumber': 'TRCK'
+	})
+
+	_v24_FIELD_MAP = frozenbidict({
+		'album': 'TALB',
+		'albumsort': 'TSOA',
+		'albumartist': 'TPE2',
+		'albumartistsort': 'TSO2',
+		'arranger': 'TPE4',
+		'artist': 'TPE1',
+		'artistsort': 'TSOP',
+		'audiodelay': 'TDLY',
+		'audiolength': 'TLEN',
+		'audiosize': 'TSIZ',
+		'bpm': 'TBPM',
+		'comment': 'COMM',
+		'compilation': 'TCMP',
+		'composer': 'TCOM',
+		'composersort': 'TSOC',
+		'conductor': 'TPE3',
+		'copyright': 'TCOP',
+		'date': 'TDRC',
 		'discnumber': 'TPOS',
 		'encodedby': 'TENC',
 		'encodingsettings': 'TSSE',
@@ -54,9 +134,7 @@ class ID3v2Frames(Tags):
 		'pictures': 'APIC',
 		'playcount': 'PCNT',
 		'publisher': 'TPUB',
-		'recordingdates': 'TRDA',
-		'subtitle': 'TSST',
-		'time': 'TIME',
+		'subtitle': 'TIT3',
 		'title': 'TIT2',
 		'titlesort': 'TSOT',
 		'tracknumber': 'TRCK'
@@ -67,15 +145,21 @@ class ID3v2Frames(Tags):
 		if not isinstance(data, DataReader):
 			data = DataReader(data)
 
-		if id3_version[1] == 2:
+		if id3_version is ID3Version.v22:
+			cls.FIELD_MAP = cls._v22_FIELD_MAP
+
 			struct_pattern = '3s3B'
 			size_len = 3
 			per_byte = 8
-		elif id3_version[1] == 3:
+		elif id3_version is ID3Version.v23:
+			cls.FIELD_MAP = cls._v23_FIELD_MAP
+
 			struct_pattern = '4s4B2B'
 			size_len = 4
 			per_byte = 8
-		elif id3_version[1] == 4:
+		elif id3_version is ID3Version.v24:
+			cls.FIELD_MAP = cls._v24_FIELD_MAP
+
 			struct_pattern = '4s4B2B'
 			size_len = 4
 			per_byte = 7
@@ -137,10 +221,10 @@ class ID3v2Header(DictMixin):
 
 		major, revision, _flags, sync_size = struct.unpack('BBB4s', data.read(7))
 
-		version = (2, major, revision)
-
-		if version[1] not in [2, 3, 4]:
-			raise ValueError("Unsupported ID3 version.")
+		try:
+			version = ID3Version((2, major))
+		except ValueError:
+			raise ValueError(f"Unsupported ID3 version (2.{major}).")
 
 		flags = DictMixin()
 
@@ -168,7 +252,7 @@ class ID3v2(DictMixin):
 
 		if self._header.flags.extended:
 			ext_size = decode_synchsafe_int(struct.unpack('4B', data.read(4))[0:4], 7)
-			if self._header.version[1] == 4:
+			if self._header is ID3Version.v24:
 				data.read(ext_size - 4)
 			else:
 				data.read(ext_size)